backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
pydantic>=2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg>=0.30.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
librosa==0.10.1
numpy>=1.26.0
torch>=2.2.0
torchaudio>=2.2.0
openai==1.3.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1

# 보컬 분리 관련
demucs>=4.0.1

# 음성인식 관련
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Google Cloud Speech-to-Text
google-cloud-speech>=2.21.0
google-auth>=2.23.0 